refactor: 상품에 대한 ViewModel 하나로 합치기
