--- app/src/main/java/woowacourse/shopping/viewmodel/ProductsViewModel.kt
@@ -1,20 +1,0 @@
-package woowacourse.shopping.viewmodel
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-import woowacourse.shopping.data.product.ProductRepository
-import woowacourse.shopping.model.Product
-
-class ProductsViewModel : ViewModel() {
-    private val _products = MutableLiveData<List<Product>>()
-    val products: LiveData<List<Product>> get() = _products
-
-    fun loadPage(
-        productRepository: ProductRepository,
-        page: Int,
-        pageSize: Int,
-    ) {
-        _products.value = productRepository.findRange(page, pageSize)
-    }
-}
